+++ packages/race-cli/src/helpers.ts
@@ -1,4 +1,5 @@
 const chalk = require('chalk');
+const commander = require('commander');
 const figlet = require('figlet');
 const {textSync} = figlet;
 
@@ -10,3 +11,12 @@ export const argumentHelper = (str: string) => chalk.green(str);
 
 export const raceLogo = `${chalk.bgBlue.bold(textSync('RacePoint'))}
 `;
+
+export const parseIntArg = (value: any) => {
+  // parseInt takes a string and a radix
+  const parsedValue = parseInt(value, 10);
+  if (isNaN(parsedValue)) {
+    throw new commander.InvalidArgumentError('Not a number.');
+  }
+  return parsedValue;
+};
